refactored timeouts to fix tests
