+++ lib/actor.js
@@ -3,8 +3,7 @@ const MetaStep = require('./step/meta')
 const recordStep = require('./step/record')
 const container = require('./container')
 const { methodsOfObject } = require('./utils')
-const { TIMEOUT_ORDER } = require('./step/timeout')
-const recorder = require('./recorder')
+const { TIMEOUT_ORDER } = require('./timeout')
 const event = require('./event')
 const store = require('./store')
 const output = require('./output')
