+++ lib/listener/globalTimeout.js
@@ -4,7 +4,7 @@ const recorder = require('../recorder')
 const Config = require('../config')
 const store = require('../store')
 const debug = require('debug')('codeceptjs:timeout')
-const { TIMEOUT_ORDER } = require('../step/timeout')
+const { TIMEOUT_ORDER, TimeoutError, TestTimeoutError } = require('../timeout')
 const { BeforeSuiteHook, AfterSuiteHook } = require('../mocha/hooks')
 
 module.exports = function () {
@@ -119,25 +119,33 @@ module.exports = function () {
     }
   })
 
+  event.dispatcher.on(event.step.after, step => {
+    if (typeof timeout !== 'number') return
+    if (!store.timeouts) return
+
+    recorder.catchWithoutStop(err => {
+      if (timeout && err instanceof TimeoutError && +Date.now() - step.startTime >= timeout) {
+        debug('Step failed due to global test or suite timeout')
+        throw new TestTimeoutError(currentTimeout)
+      }
+      throw err
+    })
+  })
+
   event.dispatcher.on(event.step.finished, step => {
     if (!store.timeouts) {
       debug('step', step.toCode().trim(), 'timeout disabled')
       return
     }
 
+    if (typeof timeout === 'number') debug('Timeout', timeout)
+
     debug(`step ${step.toCode().trim()}:${step.status} duration`, step.duration)
     if (typeof timeout === 'number' && !Number.isNaN(timeout)) timeout -= step.duration
 
     if (typeof timeout === 'number' && timeout <= 0 && recorder.isRunning()) {
       debug(`step ${step.toCode().trim()} timed out`)
-      if (currentTest && currentTest.callback) {
-        debug(`Failing test ${currentTest.title} with timeout ${currentTimeout}s`)
-        recorder.reset()
-        // replace mocha timeout with custom timeout
-        currentTest.timeout(0.1)
-        currentTest.callback(new Error(`Timeout ${currentTimeout}s exceeded (with Before hook)`))
-        currentTest.timedOut = true
-      }
+      recorder.throw(new TestTimeoutError(currentTimeout))
     }
   })
 }
