+++ lib/plugin/stepTimeout.js
@@ -1,5 +1,5 @@
 const event = require('../event')
-const { TIMEOUT_ORDER } = require('../step/timeout')
+const { TIMEOUT_ORDER } = require('../timeout')
 
 const defaultConfig = {
   timeout: 150,
