+++ lib/recorder.js
@@ -3,7 +3,7 @@ const promiseRetry = require('promise-retry')
 const chalk = require('chalk')
 const { printObjectProperties } = require('./utils')
 const { log } = require('./output')
-
+const { TimeoutError } = require('./timeout')
 const MAX_TASKS = 100
 
 let promise
@@ -386,7 +386,7 @@ function getTimeoutPromise(timeoutMs, taskName) {
   return [
     new Promise((done, reject) => {
       timer = setTimeout(() => {
-        reject(new Error(`Action ${taskName} was interrupted on step timeout ${timeoutMs}ms`))
+        reject(new TimeoutError(`Action ${taskName} was interrupted on step timeout ${timeoutMs}ms`))
       }, timeoutMs || 2e9)
     }),
     timer,
