+++ lib/step/base.js
@@ -1,6 +1,6 @@
 const color = require('chalk')
 const Secret = require('../secret')
-const { getCurrentTimeout } = require('./timeout')
+const { getCurrentTimeout } = require('../timeout')
 const { ucfirst, humanizeString, serializeError } = require('../utils')
 
 const STACK_LINE = 5
