+++ lib/step/record.js
@@ -3,7 +3,7 @@ const recorder = require('../recorder')
 const StepConfig = require('./config')
 const { debug } = require('../output')
 const store = require('../store')
-const { TIMEOUT_ORDER } = require('./timeout')
+const { TIMEOUT_ORDER } = require('../timeout')
 const retryStep = require('./retry')
 function recordStep(step, args) {
   step.status = 'queued'
