+++ lib/timeout.js
@@ -36,7 +36,23 @@ function getCurrentTimeout(timeouts) {
   return totalTimeout
 }
 
+class TimeoutError extends Error {
+  constructor(message) {
+    super(message)
+    this.name = 'TimeoutError'
+  }
+}
+
+class TestTimeoutError extends TimeoutError {
+  constructor(timeout) {
+    super(`Timeout ${timeout}s exceeded (with Before hook)`)
+    this.name = 'TestTimeoutError'
+  }
+}
+
 module.exports = {
   TIMEOUT_ORDER,
   getCurrentTimeout,
+  TimeoutError,
+  TestTimeoutError,
 }
